+++ src/components/Hero/heroChangelogData.ts
@@ -10,12 +10,22 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Styles"],
+    notes: [
+      'Updated the "campaign" variant to improve the spacing around the component.',
+      "Added the `isDarkBackgroundImage` prop.",
+    ],
+  },
+  {
     date: "2023-11-09",
     version: "2.1.2",
     type: "Update",
     affects: ["Styles"],
     notes: [
-      'Updates the layout for the "campaign" variant to have consistent padding on its left and right sides.',
+      'Updated the layout for the "campaign" variant to have consistent padding on its left and right sides.',
     ],
   },
   {